+++ app/src/main/java/de/transio/hiuni/di/DatabaseModule.kt
@@ -9,8 +9,10 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import de.transio.hiuni.core.database.AppDatabase
 import de.transio.hiuni.core.database.MIGRATION_1_2
+import de.transio.hiuni.core.database.MIGRATION_2_3
 import de.transio.hiuni.feature.calendar.data.CustomEventDao
 import de.transio.hiuni.feature.mensa.data.MealDao
+import de.transio.hiuni.feature.movies.data.MovieDao
 import javax.inject.Singleton
 
 @Module
@@ -21,7 +23,7 @@ object DatabaseModule {
     @Singleton
     fun provideAppDatabase(@ApplicationContext context: Context): AppDatabase =
         Room.databaseBuilder(context, AppDatabase::class.java, AppDatabase.DATABASE_NAME)
-            .addMigrations(MIGRATION_1_2)
+            .addMigrations(MIGRATION_1_2, MIGRATION_2_3)
             .build()
 
     @Provides
@@ -31,4 +33,8 @@ object DatabaseModule {
     @Provides
     fun provideMealDao(database: AppDatabase): MealDao =
         database.mealDao()
+
+    @Provides
+    fun provideMovieDao(database: AppDatabase): MovieDao =
+        database.movieDao()
 }